Refactor, fix priv items in pub docs, fix visibility
