+++ jason/src/api/mod.rs
@@ -22,7 +22,7 @@ use crate::{
 #[doc(inline)]
 pub use self::{
     connection::{ConnectionHandle, Connections},
-    room::Room,
+    room::{Room, RoomCloseReason},
     room::RoomHandle,
 };
 
@@ -67,18 +67,18 @@ impl Jason {
         ))))
     }
 
-    /// Returns [`RoomHandle`] for [`Room`].
+    /// Inits new Room and return [`RoomHandle`] for it.
     pub fn init_room(&self) -> RoomHandle {
         let rpc = Rc::clone(&self.0.borrow().rpc);
         self.inner_init_room(WebSocketRpcSession::new(rpc))
     }
 
-    /// Returns handle to [`MediaManager`].
+    /// Returns handle to Media Manager.
     pub fn media_manager(&self) -> MediaManagerHandle {
         self.0.borrow().media_manager.new_handle()
     }
 
-    /// Closes the provided [`Room`].
+    /// Closes the provided [`RoomHandle`].
     #[allow(clippy::needless_pass_by_value)]
     pub fn close_room(&self, room_to_delete: RoomHandle) {
         self.0.borrow_mut().rooms.retain(|room| {
@@ -102,7 +102,7 @@ impl Jason {
 }
 
 impl Jason {
-    /// Returns new [`Jason`] with the provided [`WebSocketRpcClient`].
+    /// Returns new [`Jason`] with the provided WebSocket client.
     #[inline]
     pub fn with_rpc_client(rpc: Rc<WebSocketRpcClient>) -> Self {
         Self(Rc::new(RefCell::new(Inner {
@@ -112,7 +112,7 @@ impl Jason {
         })))
     }
 
-    /// Returns [`RoomHandle`] for [`Room`].
+    /// Returns [`RoomHandle`] for Room.
     pub fn inner_init_room(&self, rpc: Rc<dyn RpcSession>) -> RoomHandle {
         let peer_repository = Box::new(peer::Repository::new(Rc::clone(
             &self.0.borrow().media_manager,
