+++ jason/src/lib.rs
@@ -4,21 +4,25 @@
 
 // TODO: Remove `clippy::must_use_candidate` once the issue below is resolved:
 //       https://github.com/rust-lang/rust-clippy/issues/4779
-#![allow(
-    clippy::module_name_repetitions,
-    clippy::must_use_candidate,
-    private_intra_doc_links
-)]
+#![allow(clippy::module_name_repetitions, clippy::must_use_candidate)]
 #![cfg_attr(not(feature = "mockable"), warn(missing_docs))]
 #![cfg_attr(feature = "mockable", allow(missing_docs))]
 
+macro_rules! export {
+    ($module:ident) => {
+        #[cfg(feature = "mockable")]
+        pub mod $module;
+        #[cfg(not(feature = "mockable"))]
+        mod $module;
+    };
+}
+
 #[macro_use]
 pub mod utils;
-
-pub mod api;
-pub mod media;
-pub mod peer;
-pub mod rpc;
+export!(media);
+export!(api);
+export!(peer);
+export!(rpc);
 
 // When the `console_error_panic_hook` feature is enabled, we can call the
 // `set_panic_hook` function at least once during initialization, and then
@@ -31,7 +35,7 @@ pub use console_error_panic_hook::set_once as set_panic_hook;
 
 #[doc(inline)]
 pub use self::{
-    api::{ConnectionHandle, Jason, RoomHandle},
+    api::{ConnectionHandle, Jason, RoomCloseReason, RoomHandle},
     media::{
         AudioTrackConstraints, DeviceVideoTrackConstraints,
         DisplayVideoTrackConstraints, FacingMode, JsMediaSourceKind, MediaKind,
