+++ src/api/client/server.rs
@@ -40,9 +40,7 @@ async fn ws_index(
 
 /// Context for [`App`] which holds all the necessary dependencies.
 pub struct Context {
-    /// Repository of all currently existing [`Room`]s in application.
-    ///
-    /// [`Room`]: crate::signalling::Room
+    /// Repository of all currently existing Rooms in application.
     pub rooms: RoomRepository,
 
     /// Settings of application.
