+++ src/api/mod.rs
@@ -22,14 +22,10 @@ pub enum RpcServerError {
     /// Authorization on the [`RpcServer`] was failed.
     Authorization,
 
-    /// [`Room`] returned some [`RoomError`].
-    ///
-    /// [`Room`]: crate::signalling::room::Room
+    /// Room returned some error.
     RoomError(RoomError),
 
-    /// [`Room`]s [`MailboxError`] is closed or overflowed.
-    ///
-    /// [`Room`]: crate::signalling::room::Room
+    /// Room's actor mailbox is closed or overflowed.
     RoomMailbox(MailboxError),
 }
 
@@ -45,11 +41,9 @@ impl From<RoomError> for RpcServerError {
 /// Server side of Medea RPC protocol.
 #[cfg_attr(test, mockall::automock)]
 pub trait RpcServer: Debug + Send {
-    /// Send signal of new [`RpcConnection`] being established with specified
-    /// [`Member`]. Transport should consider dropping connection if message
+    /// Send signal of new RPC connection being established with specified
+    /// Member. Transport should consider dropping connection if message
     /// result is err.
-    ///
-    /// [`Member`]: crate::signalling::elements::member::Member
     fn connection_established(
         &self,
         member_id: MemberId,
@@ -57,10 +51,8 @@ pub trait RpcServer: Debug + Send {
         connection: Box<dyn RpcConnection>,
     ) -> LocalBoxFuture<'static, Result<RpcConnectionSettings, RpcServerError>>;
 
-    /// Send signal of existing [`RpcConnection`] of specified [`Member`] being
+    /// Send signal of existing RPC connection of specified Member being
     /// closed.
-    ///
-    /// [`Member`]: crate::signalling::elements::member::Member
     fn connection_closed(
         &self,
         member_id: MemberId,
