+++ src/media/mod.rs
@@ -1,12 +1,10 @@
 //! Representations of media and media connection establishment objects.
 
-pub mod ice_user;
-pub mod peer;
-pub mod track;
+pub(crate) mod peer;
+pub(crate) mod track;
 
 #[doc(inline)]
 pub use self::{
-    ice_user::{IceUser, IceUsername},
     peer::{
         Peer, PeerError, PeerStateMachine, Stable, WaitLocalSdp, WaitRemoteSdp,
     },
